qcacld-3.0: Complete the wait events during recovery
Currently the event are forcefully completed during fw down
and shutdown.
During phone reboot the shutdown is not called and it may happen
that in interface down, the event is added in the link list
after receiving the fw down. So this event is not forcefully
completed and as a result of which rtnl lock is held.
IPA module waits for this rtnl lock and blocks during the shutdown
till it gets the rtnl lock.
This results in the deadlock situation and phone reboots does not
happen for 30 secs and active cmd timeout results into panic.

To mitigate this issue, complete the wait events forcefully
both in fw down and recovery uevent.

Change-Id: I265bf248fc1516a819a64e2b582c39d2fc32e760
CRs-Fixed: 2437461