qcacld-3.0: Add support for WLAN thermal mitigation
As a part of a new requirement, the WLAN subsystem needs to perform
thermal mitigation action as per the thermal state of the device. The
platform driver sends the notification with the thermal state and the
host performs the required appropriate action. The actions include
blocking northbound communication as a maximum power save state and
throttling (via duty cycle control) as an intermediate power save
action. The driver switches states and maintains the configured states
across load/unload.

Make changes to perform registration with the platform driver, act on
the thermal state as received by the platform driver and communicate
with FW regarding the mitigation action to be performed. The
communication with the FW is done via the WMI command
WMI_THERM_THROT_SET_CONF_CMDID.

Change-Id: Iedba36f83ffbdba51dfe42762efffa317d9ccb91
CRs-Fixed: 2468794