qcacld-3.0: Post scan req on priority to sme queue
If sme queue has many req pending and scan req is posted with low
priority the scan req may not get processed for long time. This
may lead to hdd_scan_inactivity_timer trigger and assert.

To fix this post the scan req to sme queue with priority so that
it will get processed before other pending commands in sme queue.

Change-Id: I503837a906ee9b14290c1cf681a83c17b699f6a7
CRs-Fixed: 2452594