qcacld-3.0: Decouple get channel command from extscan ini
Currently there is a check for the gExtscan ini value
within get valid channel command and if gExtscan ini is
disabled, get valid channel command fails.

To address above issue, do not check for the ini value
as the command to get valid channel should be independent
of ini value.

Change-Id: I6496cb94e6330f071f8027607e7d55a8f5d3db8a
CRs-Fixed: 2518313